project codes/requirements.txt
flask
flask-cors
torch
torchvision
pillow
ultralytics
opencv-python-headless
pycocotools
tqdm
pyyaml